minimal failing parse test
