+++ README.md
@@ -1,0 +1,6 @@
+# DottyDoc templating engine
+## Build
+Add the following to your `~/.sbt/0.13/global.sbt`
+```scala
+resolvers += "Artima Maven Repository" at "http://repo.artima.com/releases"
+```
