+++ app/src/antlr4/Liquid.g4
@@ -1,0 +1,30 @@
+grammar Liquid;
+
+options
+{
+language=Java;
+}
+ 
+ 
+// PARSER
+program : ((assignment|expression) ';')+;
+ 
+assignment : ID '=' expression;
+ 
+expression
+    : '(' expression ')'                # parenExpression
+    | expression ('*'|'/') expression   # multOrDiv
+    | expression ('+'|'-') expression   # addOrSubtract
+    | 'print' arg (',' arg)*            # print
+    | STRING                            # string
+    | ID                                # identifier
+    | INT                               # integer;
+ 
+arg : ID|STRING;
+ 
+// LEXER
+ 
+STRING : '"' (' '..'~')* '"';
+ID     : ('a'..'z'|'A'..'Z')+;
+INT    : '0'..'9'+;
+WS     : [ \t\n\r]+ -> skip ;
