+++ app/src/main/scala/parser.scala
@@ -1,0 +1,11 @@
+package vlthr.tee.parser
+
+import org.antlr.v4.runtime._
+import org.antlr.v4.runtime.tree._
+import vlthr.tee.core._
+
+object Parser {
+  def parseNode(node: String): Node = {
+    ???
+  }
+}
