+++ app/src/test/scala/TestMustache.scala
@@ -1,5 +1,3 @@
-import mustachemeta.Mustache
-
 object TestMustache {
   // @Mustache("template.md")
   object Template
