+++ build.sbt
@@ -2,6 +2,13 @@ lazy val metaVersion = "1.6.0"
 scalaVersion in ThisBuild := "2.12.2"
 
 
+antlr4GenListener in Antlr4 := true
+
+antlr4GenVisitor in Antlr4 := true
+
+antlr4Dependency in Antlr4 := "org.antlr" % "antlr4" % "4.5"
+
+antlr4PackageName in Antlr4 := Some("com.example")
 
 lazy val common = Seq(
   resolvers ++= Seq(
@@ -12,9 +19,9 @@ lazy val common = Seq(
 )
 
 lazy val macrosSetting = Seq(
-  scalacOptions := {
-    Seq("-Xprint:frontend,parser,macrosTransform", "-Ycheck:all")
-  },
+  // scalacOptions := {
+  //   Seq("-Xprint:frontend,parser,macrosTransform", "-Ycheck:all")
+  // },
   addCompilerPlugin("org.scalameta" % "paradise" % "3.0.0-M9" cross CrossVersion.full),
   scalacOptions += "-Xplugin-require:macroparadise",
   traceLevel := 0,
@@ -28,8 +35,10 @@ lazy val macrosSetting = Seq(
 lazy val templatingSettings = Seq(
   name := "dottydoc-templating",
   version := "0.0.1",
-  organization := "vlthr"
-) ++ common ++ macrosSetting
+  organization := "vlthr",
+  libraryDependencies += "org.scalactic" %% "scalactic" % "3.0.1" % "test",
+  libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.1" % "test"
+) ++ common ++ macrosSetting ++ antlr4Settings
 
 lazy val macros = (project in file("macros"))
   .settings(macrosSetting: _*)
