+++ project/sbt-antlr4.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("com.simplytyped" % "sbt-antlr4" % "0.7.11")
